requirement.txt
appdirs==1.4.3
Django==1.11
django-autoslug==1.9.3
django-fobi==0.10.7
django-formtools==2.0
django-nine==0.1.12
django-nonefield==0.1
easy-thumbnails==2.4.1
olefile==0.44
ordereddict==1.1
packaging==16.8
Pillow==4.1.1
pyparsing==2.2.0
pytz==2017.2
requests==2.13.0
simplejson==3.10.0
six==1.10.0
Unidecode==0.4.20
vishap==0.1.5
